set up example component
