+++ packages/fast-components/src/name-tag/fixtures/mark.html
@@ -1,0 +1,2 @@
+<fast-name-tag>Mark!</fast-name-tag>
+<fast-name-tag style="--background-color: #00F">Blue Mark!</fast-name-tag>
