+++ packages/fast-components/src/name-tag/name-tag.stories.ts
@@ -1,0 +1,15 @@
+import { FASTNameTag } from ".";
+import MarkTemplate from "./fixtures/mark.html";
+
+try {
+    // Prevent tree-shaking
+    FASTNameTag;
+} catch (e) {
+    alert(e);
+}
+
+export default {
+    title: "Name tag",
+};
+
+export const Mark = () => MarkTemplate;
