+++ packages/fast-components/src/name-tag/name-tag.template.ts
@@ -1,0 +1,16 @@
+import { html } from "@microsoft/fast-element";
+import { NameTag } from "./name-tag";
+
+export const NameTagTemplate = html<NameTag>`
+  <div class="header">
+    <slot name="avatar"></slot>
+    <h3>${x => x.greeting.toUpperCase()}</h3>
+    <h4>my name is</h4>
+  </div>
+
+  <div class="body">
+    <slot></slot>
+  </div>
+
+  <div class="footer"></div>
+`;
