+++ packages/fast-components/src/name-tag/name-tag.ts
@@ -1,0 +1,7 @@
+import { attr, FastElement } from "@microsoft/fast-element";
+
+export class NameTag extends FastElement {
+    @attr
+    public greeting: string = "Hello";
+    public bar = "";
+}
