+++ packages/fast-components/tslint.json
@@ -3,10 +3,20 @@
         "@microsoft/fast-tslint-rules",
         "tslint-config-prettier"
     ],
+    "rules": {
+        "typedef": [
+            true,
+            "call-signature",
+            "parameter",
+            "property-declaration",
+            "member-variable-declarations"
+        ]
+    },
     "linterOptions": {
         "exclude": [
             "node_modules/**",
-            "dist/**"
+            "dist/**",
+            "src/**/*.stories.ts"
         ]
     }
 }
